Commit before adding new development branch
- changed how new instance arguments work to ensure it works properly in
case OS adds it's own arguments on application start.
It will require "-i newInstance" in order for a new clean instance to
start.
This was added because I noticed KDE adds another argument at startup.
- removed Stackview from mainWindow as it was unnecessary
- changed font of the tray icon
- created a custom image provider which is used to show "Provided by
OpenWeatherMap" when their API is used, since their logo is square and
wouldn't fit properly in that space