v4.0.6
Another small commit attempting to fix issues.
This one centered on fixing a crash because of threaded access to
database, moved the initialisation of the database on application
startup.

I need to rethink how the access to database is done because SQLite does
not allow multiple connections doing multiple actions at the sametime.

Also moved the qml files into a subdirectory.